refactor: error 패키지명을 exception으로 변경
error와 exception 개념 혼란을 일으킬 수 있으므로 exception 패키지명으로 변경
