+++ src/main/java/org/example/cafe/infrastructure/database/DbConnector.java
@@ -9,7 +9,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import javax.sql.DataSource;
-import org.example.cafe.common.error.CafeException;
+import org.example.cafe.common.exception.CafeException;
 import org.slf4j.Logger;
 
 public class DbConnector {
